fix: scripts/process-scripts.py latex table builder to work with black formatted files (#114)
The file scripts/process-scripts.py processes the
ex-*.py scripts and examples directory to build latex
tables for use in the mf6examples.pdf document. The
processing of tables had an issue when the input data
contained more than 1 line.

For example:
```
   # Table TWRI Model Parameters
   nper = 1  # Number of periods
   nlay = 5  # Number of layers
```
creates a table with two rows. However the following
only makes a table with one row because the second
variable fails to load.
```
   # Table TWRI Model Parameters
   nper = 1  # Number of periods
   nlay = (
           5  # Number of layers
          )
```
which is commonly done by the python file formater `black`.

This problem was fixed by using the python module `ast`, but
would produce the record entry as
```
   Number of Layers  |   nlay = (
```
rather than
```
   Number of Layers  |   5
```

This commit fixes that isssue so that the tables now
use the appropiate line to pull the data for the latex table.